print reverse

+++ tester.sh
@@ -8,12 +8,33 @@ BIN4=./test/test2_32.out
 FILE_TEST1=./test/my.txt
 FILE_TEST2=./test/org.txt
 
+run_with_option_reverse() {
+  echo -n "      "
+  if [ "$2" = "" ]; then
+    echo -n "\"\""
+  else
+    echo -n $2
+  fi
+  echo -n " -r"
+  ./ft_nm $2 $1 "-r" > $FILE_TEST1
+  RET=$(($?))
+  nm $2 $1 "-r" > $FILE_TEST2
+  RET=$(($RET + $?))
+  diff -U 3 $FILE_TEST1 $FILE_TEST2
+  RET=$(($RET + $?))
+  if [ $RET -eq 0 ]; then
+    echo "     OK"
+  else
+    echo "     KO"
+  fi
+}
+
 run_with_option() {
   echo -n "      "
   if [ "$2" = "" ]; then
-    echo "\"\""
+    echo -n "\"\""
   else
-    echo $2
+    echo -n $2
   fi
   ./ft_nm $2 $1 > $FILE_TEST1
   RET=$(($?))
@@ -26,6 +47,7 @@ run_with_option() {
   else
     echo "        KO"
   fi
+  run_with_option_reverse $1 $2
 }
 
 do_test() {
@@ -37,7 +59,9 @@ do_test() {
   run_with_option $1 "-p"
 }
 
-rm $FILE_TEST1 $FILE_TEST2
+if test -f "$FILE_TEST1" && test -f "$FILE_TEST2"; then
+  rm $FILE_TEST1 $FILE_TEST2
+fi
 
 do_test $BIN1
 do_test $BIN2
@@ -46,4 +70,4 @@ do_test $BIN4
 do_test $OBJ
 do_test $SO
 
-rm $FILE_TEST1 $FILE_TEST2
+rm -v $FILE_TEST1 $FILE_TEST2
